backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.20
pillow==11.1.0
diffusers==0.32.1
transformers==4.47.1
accelerate==1.2.1
torch==2.5.1
torchsde==0.2.6
ultralytics==8.3.57
einops==0.8.0
segment_anything
scikit-image==0.25.0
spandrel==0.4.0